restructuring into folder SearchResults
