+++ TMdllB/SearchResults/ResultContainer.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace TMdllB
+namespace TMdllB.SearchResults
 {
     internal class ResultContainer
     {
