+++ TMdllB/TMdllB.csproj
@@ -6,4 +6,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="TV\" />
+  </ItemGroup>
+
 </Project>
